+++ cli/src/main/java/com/devonfw/tools/ide/merge/DirectoryMerger.java
@@ -2,7 +2,7 @@ package com.devonfw.tools.ide.merge;
 
 import com.devonfw.tools.ide.context.IdeContext;
 import com.devonfw.tools.ide.environment.EnvironmentVariables;
-import com.devonfw.tools.ide.merge.xmlMerger.OldXmlMerger;
+import com.devonfw.tools.ide.merge.xmlmerger.OldXmlMerger;
 import com.devonfw.tools.ide.util.FilenameUtil;
 import org.jline.utils.Log;
