+++ cli/src/main/java/com/devonfw/tools/ide/merge/xmlmerger/model/MergeElement.java
@@ -1,19 +1,24 @@
-package com.devonfw.tools.ide.merge.xmlMerger;
+package com.devonfw.tools.ide.merge.xmlmerger.model;
 
+import com.devonfw.tools.ide.merge.xmlmerger.annotation.MergeAnnotation;
 import org.w3c.dom.*;
 
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * Represents an XML element during the merge process.
+ */
 public class MergeElement {
-
   private final Element element;
 
   public MergeElement(Element element) {
+
     this.element = element;
   }
 
   public Element getElement() {
+
     return element;
   }
 
@@ -23,54 +28,65 @@ public class MergeElement {
     if (mergeStrategy != null) {
       return mergeStrategy;
     }
-    // Inherit merging strategy from parent if not explicitly specified
-    Element parent = (Element) element.getParentNode();
+
+    // Inherit merging strategy from parent
+    Element parent = getParentElement();
     if (parent != null) {
       return new MergeElement(parent).getMergingStrategy();
     }
-    return null;
+
+    return MergeStrategy.KEEP; // Default strategy
   }
 
   public String getId() {
 
-    // priority: merge:id -> parent's childrenId -> id attribute -> inherited merge:id
     String mergeId = MergeAnnotation.getMergeId(element);
     if (!mergeId.isEmpty()) {
       return mergeId;
     }
 
-    String parentChildrenId = getChildrenId();
+    String parentChildrenId = getParentMergeChildrenId();
     if (parentChildrenId != null) {
       return parentChildrenId;
     }
 
-    String idAttr = element.getAttribute("id"); // use Annotation for this
+    String idAttr = element.getAttribute("id");
     if (!idAttr.isEmpty()) {
       return idAttr;
     }
 
-    Element parent = (Element) element.getParentNode();
+    Element parent = getParentElement();
     if (parent != null) {
       return new MergeElement(parent).getId();
+    }
+
+    return null;
+  }
+
+  private String getParentMergeChildrenId() {
+
+    Element parent = getParentElement();
+    if (parent != null) {
+      String childrenId = MergeAnnotation.getMergeChildrenId(element);
+      if (!childrenId.isEmpty()) {
+        return childrenId;
+      }
+      return new MergeElement(parent).getParentMergeChildrenId();
     }
     return null;
   }
 
-  private String getChildrenId() {
+  private Element getParentElement() {
 
     Node parentNode = element.getParentNode();
-    if (parentNode == null || parentNode.getNodeType() != Node.ELEMENT_NODE) { // root's parent is document, cant be cast to elem
-      return null;
+    if (parentNode != null && parentNode.getNodeType() == Node.ELEMENT_NODE) {
+      return (Element) parentNode;
     }
-    Element parent = (Element) parentNode;
-    String childrenId = MergeAnnotation.getMergeChildrenId(element);
-    if (!childrenId.isEmpty()) {
-      return childrenId;
-    }
-    return new MergeElement(parent).getChildrenId();
+    return null;
   }
 
-  public List<MergeAttribute> getAttributes() {
+  public List<MergeAttribute> getElementAttributes() {
+    
     NamedNodeMap attributes = element.getAttributes();
     List<MergeAttribute> attributeList = new ArrayList<>();
     for (int i = 0; i < attributes.getLength(); i++) {
@@ -79,30 +95,14 @@ public class MergeElement {
     return attributeList;
   }
 
-  public List<MergeElement> getChildElements() {
-
-    NodeList childNodes = element.getChildNodes();
-    List<MergeElement> childElements = new ArrayList<>();
-    for (int i = 0; i < childNodes.getLength(); i++) {
-      if (childNodes.item(i) instanceof Element element) {
-        childElements.add(new MergeElement(element));
-      }
-    }
-    return childElements;
-  }
-
   public boolean isRootElement() {
 
-    Node parent = this.element.getParentNode();
-    if (parent != null) {
-      return parent.getNodeType() == Node.DOCUMENT_NODE;
-    }
-    return false;
+    return element.getParentNode().getNodeType() == Node.DOCUMENT_NODE;
   }
 
   public void removeMergeNSAttributes() {
 
-    List<MergeAttribute> attributes = getAttributes();
+    List<MergeAttribute> attributes = getElementAttributes();
     for (MergeAttribute attribute : attributes) {
       if (attribute.isMergeNSAttr()) {
         element.removeAttributeNode(attribute.getAttr());
@@ -122,5 +122,4 @@ public class MergeElement {
     }
     return xpath.toString();
   }
-
 }
